requirements.txt
crewai
crewai_tools
google-generativeai
selenium
webdriver-manager
requests
python-dotenv
